refactor load_data
