+++ assignment2/cs231n/load_data.py
